pipeline/requirements.txt
# Blyth Digital Twin Pipeline Dependencies

# Core geospatial
rasterio>=1.3.0
fiona>=1.9.0
shapely>=2.0.0
pyproj>=3.6.0
geopandas>=0.14.0

# Raster processing
numpy>=1.24.0
scipy>=1.11.0

# Mesh generation
trimesh>=4.0.0
pygltflib>=1.16.0

# HTTP/API
requests>=2.31.0
httpx>=0.25.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
click>=8.1.0

# Validation
pandas>=2.1.0
